started working on collecting dictionaries
