+++ backend/src/main/java/com/bakdata/conquery/io/xodus/NamespaceStorageImpl.java
@@ -4,9 +4,7 @@ import java.io.File;
 
 import javax.validation.Validator;
 
-import com.bakdata.conquery.io.xodus.stores.IdentifiableStore;
 import com.bakdata.conquery.models.config.StorageConfig;
-import com.bakdata.conquery.models.events.MasterBlockManager;
 
 import lombok.Getter;
 import lombok.NonNull;
@@ -14,13 +12,12 @@ import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
-public class NamespaceStorageImpl extends NamespacedStorageImpl<MasterBlockManager> implements NamespaceStorage {
+public class NamespaceStorageImpl extends NamespacedStorageImpl implements NamespaceStorage {
 	
 	@Getter @Setter @NonNull
 	private MasterMetaStorage metaStorage;
 	
 	public NamespaceStorageImpl(Validator validator, StorageConfig config, File directory) {
 		super(validator, config, directory);
-		this.imports = new IdentifiableStore<>(centralRegistry, StoreInfo.IMPORTS.cached(this));
 	}
 }
