+++ backend/src/main/java/com/bakdata/conquery/io/xodus/WorkerStorage.java
@@ -8,7 +8,7 @@ import javax.validation.Validator;
 
 import com.bakdata.conquery.models.config.StorageConfig;
 import com.bakdata.conquery.models.events.Block;
-import com.bakdata.conquery.models.events.SlaveBlockManager;
+import com.bakdata.conquery.models.events.BlockManager;
 import com.bakdata.conquery.models.events.CBlock;
 import com.bakdata.conquery.models.exceptions.JSONException;
 import com.bakdata.conquery.models.identifiable.ids.specific.BlockId;
@@ -18,7 +18,7 @@ import com.bakdata.conquery.models.worker.WorkerInformation;
 import jetbrains.exodus.env.Environment;
 import jetbrains.exodus.env.Environments;
 
-public interface WorkerStorage extends NamespacedStorage<SlaveBlockManager> {
+public interface WorkerStorage extends NamespacedStorage {
 	
 	WorkerInformation getWorker();
 	void setWorker(WorkerInformation worker) throws JSONException;
@@ -46,4 +46,7 @@ public interface WorkerStorage extends NamespacedStorage<SlaveBlockManager> {
 		
 		return new WorkerStorageImpl(validator, config, directory);
 	}
+	
+	void setBlockManager(BlockManager blockManager);
+	BlockManager getBlockManager();
 }
