+++ backend/src/main/java/com/bakdata/conquery/io/xodus/WorkerStorageImpl.java
@@ -9,38 +9,44 @@ import javax.validation.Validator;
 
 import com.bakdata.conquery.io.xodus.stores.IdentifiableStore;
 import com.bakdata.conquery.io.xodus.stores.SingletonStore;
+import com.bakdata.conquery.models.concepts.Concept;
 import com.bakdata.conquery.models.config.StorageConfig;
 import com.bakdata.conquery.models.datasets.Import;
 import com.bakdata.conquery.models.events.Block;
 import com.bakdata.conquery.models.events.CBlock;
-import com.bakdata.conquery.models.events.SlaveBlockManager;
+import com.bakdata.conquery.models.events.BlockManager;
 import com.bakdata.conquery.models.exceptions.ConfigurationException;
 import com.bakdata.conquery.models.exceptions.JSONException;
 import com.bakdata.conquery.models.identifiable.CentralRegistry;
 import com.bakdata.conquery.models.identifiable.ids.specific.BlockId;
 import com.bakdata.conquery.models.identifiable.ids.specific.CBlockId;
+import com.bakdata.conquery.models.identifiable.ids.specific.ConceptId;
+import com.bakdata.conquery.models.identifiable.ids.specific.ImportId;
 import com.bakdata.conquery.models.worker.WorkerInformation;
 
+import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
-public class WorkerStorageImpl extends NamespacedStorageImpl<SlaveBlockManager> implements WorkerStorage {
+public class WorkerStorageImpl extends NamespacedStorageImpl implements WorkerStorage {
 
 	private final SingletonStore<WorkerInformation> worker;
 	private final IdentifiableStore<Block> blocks;
 	private final IdentifiableStore<CBlock> cBlocks;
+	@Getter
+	private BlockManager blockManager;
 	
 	public WorkerStorageImpl(Validator validator, StorageConfig config, File directory) {
 		super(validator, config, directory);
 		this.worker = StoreInfo.WORKER.singleton(this);
-		this.imports = new IdentifiableStore<Import>(centralRegistry, StoreInfo.IMPORTS.cached(this)) {
-			@Override
-			protected void addToRegistry(CentralRegistry centralRegistry, Import imp) throws ConfigurationException, JSONException {
-				imp.loadExternalInfos(WorkerStorageImpl.this);
-			}
-		};
 		this.blocks = StoreInfo.BLOCKS.identifiable(this);
 		this.cBlocks = StoreInfo.C_BLOCKS.identifiable(this);
+	}
+	
+	@Override
+	public void setBlockManager(BlockManager blockManager) {
+		this.blockManager = blockManager;
+		blockManager.init();
 	}
 
 	@Override
@@ -116,5 +122,23 @@ public class WorkerStorageImpl extends NamespacedStorageImpl<SlaveBlockManager> 
 	@Override
 	public void updateWorker(WorkerInformation worker) throws JSONException {
 		this.worker.update(worker);
+	}
+	
+	//block manager overrides
+	@Override
+	public void updateConcept(Concept<?> concept) throws JSONException {
+		concepts.update(concept);
+		if(blockManager!=null) {
+			blockManager.removeConcept(concept.getId());
+			blockManager.addConcept(concept);
+		}
+	}
+
+	@Override
+	public void removeConcept(ConceptId id) {
+		concepts.remove(id);
+		if(blockManager!=null) {
+			blockManager.removeConcept(id);
+		}
 	}
 }
