+++ backend/src/main/java/com/bakdata/conquery/models/query/QueryPlanContext.java
@@ -1,6 +1,6 @@
 package com.bakdata.conquery.models.query;
 
-import com.bakdata.conquery.models.events.SlaveBlockManager;
+import com.bakdata.conquery.models.events.BlockManager;
 import com.bakdata.conquery.models.identifiable.CentralRegistry;
 import com.bakdata.conquery.models.worker.Worker;
 
@@ -8,7 +8,7 @@ public interface QueryPlanContext {
 
 	Worker getWorker();
 	
-	SlaveBlockManager getBlockManager();
+	BlockManager getBlockManager();
 
 	CentralRegistry getCentralRegistry();
