+++ backend/src/main/java/com/bakdata/conquery/models/query/QueryPlanContextImpl.java
@@ -1,6 +1,6 @@
 package com.bakdata.conquery.models.query;
 
-import com.bakdata.conquery.models.events.SlaveBlockManager;
+import com.bakdata.conquery.models.events.BlockManager;
 import com.bakdata.conquery.models.identifiable.CentralRegistry;
 import com.bakdata.conquery.models.worker.Worker;
 
@@ -17,7 +17,7 @@ public class QueryPlanContextImpl implements QueryPlanContext {
 	}
 
 	@Override
-	public SlaveBlockManager getBlockManager() {
+	public BlockManager getBlockManager() {
 		return worker.getStorage().getBlockManager();
 	}
 }
