+++ backend/src/main/java/com/bakdata/conquery/models/worker/Namespace.java
@@ -9,7 +9,6 @@ import java.util.concurrent.ScheduledExecutorService;
 import javax.annotation.Nonnull;
 
 import com.bakdata.conquery.io.xodus.NamespaceStorage;
-import com.bakdata.conquery.models.events.MasterBlockManager;
 import com.bakdata.conquery.models.messages.namespaces.WorkerMessage;
 import com.bakdata.conquery.models.query.QueryManager;
 import com.bakdata.conquery.models.query.entity.Entity;
@@ -32,14 +31,10 @@ public class Namespace {
 	@JsonIgnore
 	private transient Namespaces namespaces;
 	private int entityBucketSize;
-	@JsonIgnore
-	private transient MasterBlockManager blockManager;
 	
 	public Namespace(int entityBucketSize, NamespaceStorage storage) {
 		this.entityBucketSize = entityBucketSize;
 		this.storage = storage;
-		this.blockManager = new MasterBlockManager(storage);
-		this.storage.setBlockManager(blockManager);
 		this.queryManager = new QueryManager(this);
 	}
